backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.24
pydantic==2.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
